Macro bitrange::bitrange

macro_rules! bitrange {
    (
        $struct_name:ident:
        $struct_size:ty,
        $format:tt,
        $(
            $field_format:ident:
            $(
                $field_name:ident
            )+
        ),+
    ) => { ... };
}

Create a bitrange struct.

bitrange! {
    Test: u8,               // the name of the struct and the size of the internal integer
    [aaa1_0bbb],            // the format of the bits in the internal integer
    a: first,               // map the bits that are marked as `a` to field `first`
    b: second set_second    // map the bits that are marked as `b` to field `second`
                            // and create a setter `set_second` that sets a given value to `b`
}